Rename component directory back to Mirror (instead of Sync)
