+++ ui/src/pages/namespace/Mirror/Mirror/Detail/ActivityDetail.tsx
