+++ ui/src/pages/namespace/Mirror/Mirror/Detail/Header/index.tsx
