+++ ui/src/pages/namespace/Mirror/Mirror/Detail/Logs/Entry.tsx
