+++ ui/src/pages/namespace/Mirror/Mirror/Detail/Logs/index.tsx
