+++ ui/src/pages/namespace/Mirror/Mirror/Detail/index.tsx
