+++ ui/src/pages/namespace/Mirror/Mirror/Header/SyncDialog.tsx
