+++ ui/src/pages/namespace/Mirror/Mirror/Row/index.tsx
