+++ ui/src/pages/namespace/Mirror/Mirror/index.tsx
